feat: dropdown overlay is updated to dropdown
